remove type casting

+++ mlx_vlm/models/qwen2_vl/qwen2_vl.py
@@ -78,7 +78,7 @@ class Model(nn.Module):
         # Positions of <image> tokens in input_ids, assuming batch size is 1
         image_positions = input_ids == image_token_index
         image_indices = np.where(image_positions)[1].tolist()
-        inputs_embeds[:, image_indices, :] = image_features.astype(mx.float32)
+        inputs_embeds[:, image_indices, :] = image_features
 
         # TODO: Add video features
